+++ tests/test_yacman_alias.py
@@ -1,5 +1,6 @@
 import pytest
 import yacman
+from yacman.exceptions import AliasError
 
 
 class TestAliases:
@@ -16,6 +17,79 @@ class TestAliases:
         key = list(entries.keys())[0]
         alias = list(aliases.values())[0][0]
         assert x[key] == x[alias]
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_aliases_init_valid_fun(self, entries, aliases):
+        print(f"aliases: {aliases}")
+
+        def _aliases_fun(x):
+            return aliases
+
+        yacman.AliasedYacAttMap(entries=entries, aliases=_aliases_fun)
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_aliases_init_invalid_fun(self, entries, aliases):
+        print(f"aliases: {aliases}")
+
+        def _aliases_fun():
+            return aliases
+
+        with pytest.raises(AliasError):
+            yacman.AliasedYacAttMap(
+                entries=entries, aliases=_aliases_fun, aliases_strict=True
+            )
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_aliases_init_fun_error(self, entries, aliases):
+        print(f"aliases: {aliases}")
+
+        def _aliases_fun():
+            raise Exception("test")
+
+        with pytest.raises(AliasError):
+            yacman.AliasedYacAttMap(
+                entries=entries, aliases=_aliases_fun, aliases_strict=True
+            )
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_aliases_init_fun_returns_invalid(self, entries, aliases):
+        print(f"aliases: {aliases}")
+
+        def _aliases_fun():
+            return {"a": "test"}
+
+        with pytest.raises(AliasError):
+            yacman.AliasedYacAttMap(
+                entries=entries, aliases=_aliases_fun, aliases_strict=True
+            )
 
     @pytest.mark.parametrize(
         ["entries", "aliases"],
@@ -152,3 +226,40 @@ class TestAliases:
         assert x.set_aliases(key=key, aliases=alias)
         assert key in x
         assert alias[0] in x
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_delitem_checks_for_keys_and_aliases(self, entries, aliases):
+        print(f"aliases: {aliases}")
+        x = yacman.AliasedYacAttMap(entries=entries)
+        key = list(aliases.keys())[0]
+        alias = aliases[key]
+        x.__delitem__(key)
+        assert key not in x
+        x = yacman.AliasedYacAttMap(entries=entries)
+        x.set_aliases(key=key, aliases=alias)
+        x.__delitem__(alias[0])
+        assert key not in x
+
+    @pytest.mark.parametrize(
+        ["entries", "aliases"],
+        [
+            ({"a": "1"}, {"a": ["alias_a"]}),
+            ({"b": "1"}, {"b": ["alias_b"]}),
+            ({"55": "1"}, {"55": ["alias_55"]}),
+        ],
+    )
+    def test_containment_correctly_resolves_to_false(self, entries, aliases):
+        print(f"aliases: {aliases}")
+        x = yacman.AliasedYacAttMap(entries=entries)
+        key = list(aliases.keys())[0]
+        alias = aliases[key]
+        assert x.set_aliases(key=key, aliases=alias)
+        assert not (f"{key}_false" in x)
+        assert not (f"{alias[0]}_false" in x)