+++ yacman/const.py
@@ -2,15 +2,20 @@
 Constant variables for yacman package
 """
 
-FILEPATH_KEY = "_file_path"
-RO_KEY = "_ro"
-USE_LOCKS_KEY = "_locks"
-ORI_STATE_KEY = "_ori_state"
-WAIT_MAX_KEY = "_wait_time"
-ALIASES_KEY = "__aliases"
-ALIASES_KEY_RAW = "__aliases_raw"
+IK = "__internal"
+
+FILEPATH_KEY = "file_path"
+RO_KEY = "ro"
+USE_LOCKS_KEY = "locks"
+ORI_STATE_KEY = "ori_state"
+WAIT_MAX_KEY = "wait_time"
+ALIASES_KEY = "aliases"
+ALIASES_KEY_RAW = "aliases_raw"
+WRITE_VALIDATE_KEY = "write_validate"
+SCHEMA_KEY = "schema"
 
 ATTR_KEYS = (
+    IK,
     USE_LOCKS_KEY,
     FILEPATH_KEY,
     RO_KEY,
@@ -18,6 +23,8 @@ ATTR_KEYS = (
     WAIT_MAX_KEY,
     ALIASES_KEY,
     ALIASES_KEY_RAW,
+    WRITE_VALIDATE_KEY,
+    SCHEMA_KEY,
 )
 
 LOCK_PREFIX = "lock."